Dogs_Cats_CNN/requirements.txt
streamlit==0.67.1
numpy==1.19.2
tensorflow-cpu==2.1.0
Pillow==6.2.0